% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbinderAPI.R
\name{binder_andrea}
\alias{binder_andrea}
\title{Run the binder algorithm using Andrea implementation (TBD)

TBD}
\usage{
binder_andrea(CI, weight = 0.5)
}
\arguments{
\item{C}{fit CI}

\item{Const_Binder}{Const_Binder}
}
\description{
Run the binder algorithm using Andrea implementation (TBD)

TBD
}
