% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbinderAPI.R
\name{binder_naive}
\alias{binder_naive}
\title{Run the naive binder algorithm (TBD)

TBD}
\usage{
binder_naive(CI, Const_Binder = 0.5)
}
\arguments{
\item{Const_Binder}{Const_Binder}

\item{C}{fit CI}
}
\description{
Run the naive binder algorithm (TBD)

TBD
}
