% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbinderAPI.R
\name{binder_opt}
\alias{binder_opt}
\title{Run the ptimized binder algorithm (TBD)

TBD}
\usage{
binder_opt(CI, Const_Binder = 0.5)
}
\arguments{
\item{Const_Binder}{Const_Binder}

\item{C}{fit CI}
}
\description{
Run the ptimized binder algorithm (TBD)

TBD
}
