% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbinderAPI.R
\name{binder_parallel}
\alias{binder_parallel}
\title{Run the parallel binder algorithm (TBD)

TBD}
\usage{
binder_parallel(CI, Const_Binder = 0.5)
}
\arguments{
\item{Const_Binder}{Const_Binder}

\item{C}{fit CI}
}
\description{
Run the parallel binder algorithm (TBD)

TBD
}
